backend/requirements.txt
agent-starter-pack==0.8.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
claude-code-sdk==0.0.14
click==8.2.1
cookiecutter==2.6.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
fastapi==0.115.14
file-magic==0.4.1
git-filter-repo==2.47.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-documentai==3.5.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.24.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.0
openai==1.93.3
packaging==25.0
pandas==2.3.0
postgrest==1.1.1
proto-plus==1.26.1
protobuf==5.29.5
psutil
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.7.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.3
requests==2.32.4
rich==14.0.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tenacity==8.5.0
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
pdfplumber==0.11.4
PyPDF2
PyMuPDF==1.24.14
pytesseract==0.3.13
Pillow==11.1.0
boto3>=1.26.137
botocore>=1.29.137
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1
transformers==4.35.0
pytest-asyncio==0.22.0
